pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "murmur-agent"
version = "1.0.0"
description = "A coordinated swarm of AI agents for your codebase."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = [
    "ai", "agents", "coding", "cli", "langgraph",
    "mcp", "llm", "swarm", "automation", "murmur"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "langgraph>=0.3.0",
    "langgraph-supervisor>=0.0.18",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-nvidia-ai-endpoints>=0.3.0",
    "mcp>=1.0.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "filelock>=3.14.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "anyio>=4.4.0",
    "pathspec>=0.12.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "aiosqlite>=0.20.0",
    "gitpython>=3.1.40",
    "chromadb>=0.5.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
google = ["langchain-google-genai>=2.0.0"]
groq = ["langchain-groq>=0.2.0"]
faiss = ["faiss-cpu>=1.8.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "mypy>=1.10.0",
]

[project.scripts]
murmur = "murmur.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["murmur*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
